Reduce PWDN delay from 1000 to 20 ms
It took 1s to open the devive because of the delay here. A delay of 20ms
should be enough for the chip to wake up from deep shutdown and react to
the RESET signal that follows.

Also assert the reset to "high" before power-up, to prevent the reset
being low still.