Assert the RESET pin only when performing a reset
The driver kept the RESET line low (asserted) when the chip was not
in use. This was not how the line is supposed to be used. Keep the
reset signal high at all times, and only assert it shortly to pulse
a reset command to the chip(s). Unlike some other chips, asserting
RESET will not result in power savings. In fact, it may interfere
with correct operation.

The driver already uses the "standby" command to put the chip in low
power mode when not in use.